Moving the welcome func/handler
The welcome function gets set up on the root of the web application. That is
fine if this is being run as a stand alone application. If it is being
incorporated into another application. This moves the function and handler into
the main, so it won't be present when used as a library.