Simple implementation of search
This code is the start of an implementation of search for resources specific to
the patient matching test harness. The current code allows for search against
RecordMatchRuns by recordMatchContextId. It is possible to extend this
implementation by adding to the SearchParams variable in controllers.

The current implementation infers the type of the search parameter by name. If
the name ends in "Id", it assumes that the type should be bson.ObjectId.
Otherwise, it is a string.